requirements.txt
absl-py==2.1.0
accelerate==0.29.3
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.1.0
bitsandbytes==0.42.0
bleach==6.1.0
boto3==1.28.56
botocore==1.31.56
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
comm==0.2.2
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
Django==4.2.4
django-cors-headers==4.2.0
django-dotenv==1.4.2
django-ses==3.5.0
django-storages==1.14.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ecdsa==0.18.0
evaluate==0.4.1
executing==2.0.1
Faker==24.4.0
fastapi==0.110.0
fastapi-pagination==0.12.23
fastjsonschema==2.19.1
filelock==3.13.4
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2024.2.0
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.0
grpcio==1.62.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
inquirerpy==0.3.4
install==1.3.5
ipykernel==6.29.4
ipython==8.23.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
keras==3.2.1
kombu==5.3.2
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.16
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==23.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pfzy==0.3.4
Pillow==10.0.0
platformdirs==4.2.0
pluggy==1.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.2.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2023.3
PyYAML==6.0.1
pyzmq==26.0.0
redis==5.0.1
referencing==0.34.0
regex==2024.4.16
reportlab==4.0.4
requests==2.31.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.7.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
Send2Trash==1.8.3
sendgrid==6.10.0
sentence-transformers==2.7.0
sentry-sdk==1.29.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.22
sqlparse==0.4.4
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.36.3
strgen==1.3.1
StringGenerator==0.4.4
stripe==7.12.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.4.0
tinycss2==1.2.1
tokenizers==0.15.2
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
twilio==8.5.0
Twisted==23.8.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.29.0
uWSGI==2.0.22
vine==5.0.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
widgetsnbextension==3.6.6
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.2
zope.interface==6.0
